Add ability to pre-load SSH connections with #pre_load_ssh
Normally, Netid will only initialize a SSH connection when a command is recieved to be executed on that specific host (lazy loading). Once it establishes a connection, it adds it to the connection pool to be re-used if required.

However, this is a blocking operation: it cannot execute a command until SSH is established. This is normally okay for applications with one or two connections, but this waiting can stack up (someting like 200ms per additional host).

The instance method #pre_load_ssh will initialize the connection pool with the specified hosts. These connections are threaded and will run in parallel, saving around (n-1)*200ms in overall execution where these threads would have already been established)

This is an experimental feature